Add whoami module and continue working on permissions module

+++ functions/data_parser.py
@@ -43,9 +43,10 @@ class Parser:
         data_category = executed_module.split("_")[0]
 
         parsed_data = {}
-        parsed_data[data_category] = module_results
+        parsed_data[data_category] = []
+        parsed_data[data_category].append(module_results)
 
-        return parsed_data 
+        return parsed_data
 
     def store_parsed_results(self, selected_session, parsed_results):
         category = list(parsed_results)[0]
@@ -53,6 +54,22 @@ class Parser:
         if not os.path.exists(category_path):
             os.mkdir(category_path)
         
-        results_file = open("{}/{}_results.json".format(category_path, category), "w")
-        json.dump(parsed_results, results_file, default=str)
-        results_file.close()
+        results_file_path = "{}/{}_results.json".format(category_path, category)
+
+        if os.path.exists(results_file_path):
+            results_file = open(results_file_path, "r")
+            existent_data = json.load(results_file)
+            results_file.close()
+
+            for result in parsed_results[category]:
+                if result in existent_data[category]:
+                    pass
+                else:
+                    existent_data[category].append(result)
+
+            results_file = open(results_file_path, "w")
+            json.dump(existent_data, results_file, default=str)
+        else:
+            results_file = open("{}/{}_results.json".format(category_path, category), "w+")
+            json.dump(parsed_results, results_file, default=str)
+            results_file.close()
